+++ just_audio/CHANGELOG.md
@@ -7,6 +7,21 @@
 
 * Null safety.
 
+## 0.6.14+1
+
+* Update ICY metadata feature status in README.
+
+## 0.6.14
+
+* Initial support for ICY metadata on iOS.
+* Upgrade to ExoPlayer 2.13.1 (@MichealReed).
+
+## 0.6.13
+
+* Add userAgent parameter to AudioPlayer constructor.
+* Support seeks during download in LockCachingAudioSource.
+* Fix negative duration bug on iOS/macOS.
+
 ## 0.6.12
 
 * Fix non-completing seek-to-index bug on iOS/macOS.